Create scoped proxy for ApplicationEvents
As a proof of concept, this commit introduces scoped proxy support for
an ApplicationEvents bean that is managed by the SimpleThreadScope.

Although this approach works in general (in terms of providing an
ApplicationEvents bean that can be autowired into test classes and
retrieved as a bean by third-party extensions), it has a major
drawback. Specifically, an ApplicationEvents instance will be created
on-demand for the current thread whenever the
ThreadBoundApplicationListener attempts to access the
ApplicationEvents, and this will occur even if the user has not
annotated the test class with @RecordApplicationEvents.

In light of the above, this commit will likely be subsequently reverted
but remains in tact in the commit history for possible future
consideration.

See gh-25616