Support parallel test execution for ApplicationEvents
Prior to this commit, if multiple tests were being executed in parallel
for a test class with per-test-method test instance lifecycle semantics
(e.g., in JUnit 4 and the default in JUnit Jupiter), then a race
condition could occur when registering the
ApplicationEventsApplicationListener in the ApplicationContext.
Consequently, multiple instances of ApplicationEventsApplicationListener
would be registered resulting in duplicate tracked events in
ApplicationEvents.

This commit avoids this race condition via a synchronized block in
ApplicationEventsTestExecutionListener's
registerListenerAndResolvableDependencyIfNecessary() method.

See gh-25616