Introduce "search enclosing class" predicate for MergedAnnotations
This commit is a "proof of concept" for introducing a
`Predicate<Class<?>> searchEnclosingClass` for use with the
TYPE_HIERARCHY search strategy as a replacement for the deprecated
TYPE_HIERARCHY_AND_ENCLOSING_CLASSES search strategy.

Currently (in this commit) it is possible for a user to supply a
searchEnclosingClass predicate for any SearchStrategy; however, the
predicate is only honored if the search strategy is TYPE_HIERARCHY.
Thus, we may later choose to throw an exception is the user supplies a
predicate for a search strategy other than TYPE_HIERARCHY.

This commit also demonstrates that this new feature can be utilized in
TestContextAnnotationUtils for the implementation of hasAnnotation(...)
and findMergedAnnotation(...).

See gh-28080