Introduce fluent API for MergedAnnotations searches
This commit is a "proof of concept" for introducing a fluent API for
MergedAnnotations searches.

This commit replaces the SearchAlgorithm from the previous commit by
adding a builder directly to MergedAnnotations and introducing static
`find*(...)` factory methods in MergedAnnotations such as
findDirectlyDeclaredAnnotations() and other similar methods matching
the available strategies in SearchStrategy.

A user can only supply a searchEnclosingClass predicate for
SearchStrategy.TYPE_HIERARCHY via
MergedAnnotations.findAnnotationsInTypeHierarchy(Predicate<Class<?>>).

This leads to usage such as the following in TestContextAnnotationUtils.

MergedAnnotations
	.findAnnotationsInTypeHierarchy(TestContextAnnotationUtils::searchEnclosingClass)
	.from(clazz)
	.isPresent(annotationType);

See gh-28080