Allow collection creation from var-args in CollectionFactory
This commit serves as a proof of concept for factory methods in
CollectionFactory that allow a collection to be created as a one-liner
using var-args.

createCollection(Class<C>, Class<E>, E...) internally delegates to
createCollection(Class<?>, Class<?>, int) to create the collection
based on the supplied collectionType and populates the collection from
the var-args. The elementType is only required when creating an EnumSet
and could therefore conceivably be omitted if we do not want to support
EnumSet for this collection factory method.

collectionOf(Supplier<C>, E...) creates the collection using the
supplied collectionFactory and populates the collection from the
var-args.

See gh-28025