Introduce support for forking a GenerationContext for a new target class
This commit reverts the changes to ApplicationContextAotGenerator and
ApplicationContextInitializationCodeGenerator and replaces that
functionality with support for forking a GenerationContext for a new
target class.

This allows GeneratedClasses to contain all generated classes by
transparently changing the current ClassNameGenerator behind the scenes.
The state of the ClassNameGenerator is also retained transparently.

See gh-28928