Introduce TestAotPropertiesAware callback
This allows test components that need to work with TestAotProperties
to opt into such support by implementing TestAotPropertiesAware.

This also removes the need for a static instance of TestAotProperties
and simplifies test AOT properties handling within
TestContextAotGenerator.

One drawback to this approach is that TestAotProperties support is now
limited to test components that implement TestAotPropertiesAware, and
currently only the following types of test components are supported.

- org.springframework.test.context.ContextLoader
- org.springframework.test.context.TestContextBootstrapper
- org.springframework.test.context.TestExecutionListener