Restore classpath scanning support for Windows filesystem
The recent switch to java.nio.file.FileSystem-based classpath scanning
(see gh-29163) resulted in a regression when scanning the classpath in
the Windows filesystem.

This commit addresses this regression by introducing several
workarounds for dealing with `file:` resources on Windows.

See the in-line comments in this commit for details on those
workarounds.

This fix has been tested in the following environments.

- macOS 12.6 / Java 17 Oracle (build 17.0.4.1+1-LTS-2)

- macOS 12.6 / GraalVM CE 22.2.0 (build 17.0.4+8-jvmci-22.2-b06)

- MS Windows 11 / Java 17 Microsoft (build 17.0.4.1+1-LTS)

Closes gh-29226