Revise contribution
This commit partially reverts changes to SpelNodeImpl in order to
reduce the scope of the overall change set.

In addition, this commit revises the previous commit to support
type-safe checks for array subtype compatibility.

In order to support backward compatibility, this commit also reintroduces
generateCodeForArguments(MethodVisitor, CodeFlow, Member, SpelNodeImpl[])
in deprecated form.

Closes gh-32804