Introduce DynamicPropertyRegistrar support
This is a PROOF OF CONCEPT which introduces DynamicPropertyRegistrar
support as a replacement for registering a DynamicPropertyRegistry as
a bean in the ApplicationContext.

It appears that most things work as expected; however, in this POC there
are two instances of DefaultDynamicPropertyRegistry that prevent the use
of @⁠DynamicPropertySource and DynamicPropertyRegistrar in the same
ApplicationContext, since the two features end up writing to two different
valueSuppliers maps. The commented-out "magic.word" assertions in
DynamicPropertySourceIntegrationTests therefore currently fail.

See gh-33501