Revise DynamicPropertyRegistrar support
With this commit, a DynamicValuesPropertySource is now created and
registered with the Environment on demand whenever a
DefaultDynamicPropertyRegistry is created, which only occurs once in
DynamicPropertiesContextCustomizer and once in
DynamicPropertyRegistrarBeanInitializer.

In addition, DefaultDynamicPropertyRegistry now operates on the
valueSuppliers map stored in the singleton DynamicValuesPropertySource
registered with the Environment, which allows @⁠DynamicPropertySource
methods and DynamicPropertyRegistrar beans to transparently populate
the same DynamicValuesPropertySource.

See gh-33501