Register test bean override infrastructure beans as singletons
Prior to this commit, AOT processing failed for tests that made use of
test bean override features, since the Set<OverrideMetadata>
constructor argument configured in the bean definition for the
BeanOverrideBeanFactoryPostProcessor could not be handled by our AOT
support.

To address that, this commit registers test bean override
infrastructure beans as singletons instead of via bean definitions with
the infrastructure role.

See gh-32933