Create real bean definitions for test bean overrides
Prior to this commit, test bean overrides were always created
programmatically and registered as singletons in the bean factory,
alongside "dummy" bean definitions; however, that can lead to issues
during AOT processing and in the AOT runtime, especially for Mockito
based mocks and spies since the "dummy" bean definitions are not real
bean definitions that can be used to create a bean instance in the bean
factory.

To address that, this commit introduces a new
updateOverrideBeanDefinition(BeanDefinition) method in the
OverrideMetadata SPI that allows a concrete OverrideMetadata
implementation to update the BeanDefinition so that it is suitable for
use in AOT -- for example, by registering a static factory method that
is responsible for creating the bean instance.

This commit updates TestBeanOverrideMetadata and DummyOverrideMetadata
in order to verify that this approach could potentially work for
@⁠TestBean and custom bean override implementations. However, this
commit does not attempt to convert the @⁠MockitoBean or
@⁠MockitoSpyBean support to use this approach.

In other words, this commit serves merely as a proof of concept for an
approach that we could consider.

See gh-32933