Pause inactive application contexts in the TestContext framework
Since the introduction of the Spring TestContext Framework in 2007,
application contexts have always been stored in the context cache in an
"active" state. However, leaving a context in an active state means
that components in the context may continue to run in the background.
For example, JMS listeners may continue to consume messages from a
queue; scheduled tasks may continue to perform active work, etc.; and
this can lead to issues within a test suite.

To address such use cases, this commit introduces built-in support for
pausing inactive application contexts when they are not in use and
restarting them when they are needed again.

Specifically, a new ContextStateTestExecutionListener marks a test's
application context as "inactive" after execution of the test class has
ended, and the underlying ContextCache then "stops" the application
context. When a TestExecutionListener later attempts to obtain the
application context -- for example, for a test class that shares the
same application context-- the ContextCache will "restart" any paused
application context before returning it.

See https://github.com/spring-projects/spring-boot/issues/28312
See gh-35171
Closes gh-35168