Support POJO programming model for TestNG in the TestContext framework
This commit is merely a Proof Of Concept, and although it makes
substantial progress toward the end goal, it is unfortunately not fully
functional.

The reason is that TestNG 7.11 still does not provide suitable
extension APIs or mechanisms that would allow the Spring TestContext
Framework to be fully supported in a POJO programming model.

For example, unless I have overlooked something, it is currently
impossible to register a listener with TestNG that will be invoked just
prior to when any @⁠BeforeMethod methods _would_ be invoked. You can
implement org.testng.IInvokedMethodListener's beforeInvocation(...)
callback, but that will only be invoked prior to a @⁠BeforeMethod or
@⁠Test method. Whereas, we need a callback that will *always* be
invoked prior to the point in time that @⁠BeforeMethod methods would be
invoked, even if there are not any @⁠BeforeMethod methods in the
current test class.

Plus, if beforeInvocation(...) is invoked for a @⁠BeforeMethod method,
it does not appear to be possible to obtain a reference to the @⁠Test
method which will subsequently be invoked, which Spring needs in order
to look up annotations on the @⁠Test method (such as @⁠Transactional,
etc.).

Furthermore, the test instance supplied to TestContextManager's
prepareTestInstance() callback had to be looked up via a bit of a hack
in SpringListener.onBeforeClass(ITestClass). Whereas, we would prefer
to be able to implement a TestNG listener callback that gets invoked
directly with the test instance, immediately after TestNG has
instantiated the test class.

Last but not least, when TestNG listeners (like the SpringListener in
this commit) are registered via @⁠Listeners directly on a test class,
those listeners are actually registered once, globally for the entire
TestNG test suite. Whereas, we would prefer that the SpringListener
only be applied to test classes annotated (or meta-annotated) with
@⁠Listeners(SpringListener.class).

See gh-17422